+++ src/test/emulators/auth/helpers.ts
@@ -134,6 +134,26 @@ export async function signInWithEmailLink(
     });
 }
 
+export function signInWithPassword(
+  testAgent: TestAgent,
+  email: string,
+  password: string
+): Promise<{ idToken: string; localId: string; refreshToken: string; email: string }> {
+  return testAgent
+    .post("/identitytoolkit.googleapis.com/v1/accounts:signInWithPassword")
+    .send({ email, password })
+    .query({ key: "fake-api-key" })
+    .then((res) => {
+      expectStatusCode(200, res);
+      return {
+        idToken: res.body.idToken,
+        localId: res.body.localId,
+        refreshToken: res.body.refreshToken,
+        email: res.body.email,
+      };
+    });
+}
+
 export async function signInWithPhoneNumber(
   testAgent: TestAgent,
   phoneNumber: string
@@ -227,7 +247,19 @@ export function getAccountInfoByIdToken(testAgent: TestAgent, idToken: string): 
     .query({ key: "fake-api-key" })
     .then((res) => {
       expectStatusCode(200, res);
-      expect(res.body.users).to.have.length(1);
+      expect(res.body.users || []).to.have.length(1);
+      return res.body.users[0];
+    });
+}
+
+export function getAccountInfoByLocalId(testAgent: TestAgent, localId: string): Promise<UserInfo> {
+  return testAgent
+    .post("/identitytoolkit.googleapis.com/v1/accounts:lookup")
+    .send({ localId: [localId] })
+    .set("Authorization", "Bearer owner")
+    .then((res) => {
+      expectStatusCode(200, res);
+      expect(res.body.users || []).to.have.length(1);
       return res.body.users[0];
     });
 }